git-svn-id: svn://192.168.2.130:8807/sd/java/financialSystem@51772 ced6f03c-c1bb-49ab-ae0c-4da96410de7b

+++ src/main/java/cn/financial/service/impl/OrganizationServiceImpl.java
@@ -39,23 +39,27 @@ public class OrganizationServiceImpl implements OrganizationService {
         Map<Object, Object> map = new HashMap<Object, Object>();
         map.put("code", organization.getParentId());
         // 找到父节点信息
-        Organization org = organizationDAO.getOrganizationById(parentOrgId);
-        if (null != org) {
+        Map<Object, Object> mapparentOrgId = new HashMap<>();
+        mapparentOrgId.put("id", parentOrgId);
+        List<Organization> org = organizationDAO.listOrganizationBy(mapparentOrgId);
+        if (!CollectionUtils.isEmpty(org)) {
             // 找到兄弟节点信息
-            List<Organization> list = organizationDAO.listByParentId(org.getCode());
+            Map<Object, Object> mapbro = new HashMap<>();
+            mapbro.put("parentId", org.get(0).getCode());
+            List<Organization> list = organizationDAO.listOrganizationBy(mapbro);
             // 若存在兄弟节点，则兄弟节点的code找到该节点的code
-            if (null != list && list.size() != 0) {
+            if (!CollectionUtils.isEmpty(list)) {
                 List<String> codes = new ArrayList<String>();
                 for (Organization orga : list) {
                     codes.add(orga.getCode());
                 }
-                code = UuidUtil.getCodeByBrother(org.getCode(), codes);
+                code = UuidUtil.getCodeByBrother(org.get(0).getCode(), codes);
             }
             // 若不存在兄弟节点，那code就是父节点的code加上01
             else {
-                code = org.getCode() + "01";
+                code = org.get(0).getCode() + "01";
             }
-            organization.setParentId(org.getCode());// 父id
+            organization.setParentId(org.get(0).getCode());// 父id
             organization.setCode(code); // 该组织机构节点的序号，两位的，比如（01；0101，0102）
             organization.setHis_permission(code); // 新增时，历史权限id就是此节点的code
             return organizationDAO.saveOrganization(organization);
@@ -65,24 +69,10 @@ public class OrganizationServiceImpl implements OrganizationService {
     }
 
     /**
-     * 查询所有的组织结构
-     */
-    public List<Organization> listOrganization() {
-        return organizationDAO.listOrganization();
-    }
-
-    /**
      * 根据条件查询组织结构信息
      */
     public List<Organization> listOrganizationBy(Map<Object, Object> map) {
         return organizationDAO.listOrganizationBy(map);
-    }
-
-    /**
-     * 根据ID查询组织结构信息
-     */
-    public Organization getOrganizationById(String id) {
-        return organizationDAO.getOrganizationById(id);
     }
 
     /**
@@ -100,25 +90,29 @@ public class OrganizationServiceImpl implements OrganizationService {
     }
 
     /**
-     * 伪删除 <根据组织结构ID修改状态为0，即已删除>,停用(单条停用)
+     * 停用(单条停用),根据组织结构ID修改状态为0，即已停用
      */
     public Integer deleteOrganizationByStatus(String id) {
         return organizationDAO.deleteOrganizationByStatus(id);
     }
 
     /**
-     * 伪删除 <根据组织结构ID修改状态为0，即已删除>,停用(级联停用，将此节点下的所有子节点停用)
+     * 停用(级联停用，将此节点下的所有子节点停用)，根据组织结构ID修改状态为0，即已停用
      */
     public Integer deleteOrganizationByStatusCascade(String id) {
-        // 根据id查询到该节点信息
-        Organization org = organizationDAO.getOrganizationById(id);
-        // 根据该节点的code查询到其下所有子节点信息的集合
-        List<Organization> list = organizationDAO.listTreeByCodeForSon(org.getCode());
         Integer i = 0;
-        Iterator<Organization> iterator = list.iterator();
-        while (iterator.hasNext()) {
-            Organization orga = iterator.next();
-            i = organizationDAO.deleteOrganizationByStatus(orga.getId());
+        // 根据id查询到该节点信息
+        Map<Object, Object> map = new HashMap<>();
+        map.put("id", id);
+        List<Organization> org = organizationDAO.listOrganizationBy(map);
+        if (!CollectionUtils.isEmpty(org)) {
+            // 根据该节点的code查询到其下所有子节点信息的集合
+            List<Organization> list = organizationDAO.listTreeByCodeForSon(org.get(0).getCode());
+            Iterator<Organization> iterator = list.iterator();
+            while (iterator.hasNext()) {
+                Organization orga = iterator.next();
+                i = organizationDAO.deleteOrganizationByStatus(orga.getId());
+            }
         }
         return i;
     }
@@ -128,9 +122,12 @@ public class OrganizationServiceImpl implements OrganizationService {
      */
     @Override
     public String TreeByIdForSon(String id) {
-        Organization organizationByIds = organizationDAO.getOrganizationById(id);
-        if (null != organizationByIds) {
-            List<Organization> list = organizationDAO.listTreeByCodeForSon(organizationByIds.getCode());
+        // 根据id查询到该节点信息
+        Map<Object, Object> map = new HashMap<>();
+        map.put("id", id);
+        List<Organization> organizationByIds = organizationDAO.listOrganizationBy(map);
+        if (CollectionUtils.isEmpty(organizationByIds)) {
+            List<Organization> list = organizationDAO.listTreeByCodeForSon(organizationByIds.get(0).getCode());
             if (!CollectionUtils.isEmpty(list)) {
                 List<TreeNode<Organization>> nodes = new ArrayList<>();
                 String jsonStr = "";
@@ -151,21 +148,16 @@ public class OrganizationServiceImpl implements OrganizationService {
     }
 
     /**
-     * 根据parentid查询节点信息
-     */
-    @Override
-    public List<Organization> listByParentId(String parentId) {
-        return organizationDAO.listByParentId(parentId);
-    }
-
-    /**
      * 根据id查询该节点下的所有子节点集合
      */
     @Override
     public List<Organization> listTreeByIdForSon(String id) {
-        Organization organizationByIds = organizationDAO.getOrganizationById(id);
-        if (null != organizationByIds) {
-            List<Organization> list = organizationDAO.listTreeByCodeForSon(organizationByIds.getCode());
+        // 根据id查询到该节点信息
+        Map<Object, Object> map = new HashMap<>();
+        map.put("id", id);
+        List<Organization> organizationByIds = organizationDAO.listOrganizationBy(map);
+        if (!CollectionUtils.isEmpty(organizationByIds)) {
+            List<Organization> list = organizationDAO.listTreeByCodeForSon(organizationByIds.get(0).getCode());
             if (!CollectionUtils.isEmpty(list)) {
                 return list;
             }
@@ -178,9 +170,12 @@ public class OrganizationServiceImpl implements OrganizationService {
      */
     @Override
     public List<Organization> listTreeByIdForParent(String id) {
-        Organization organizationByIds = organizationDAO.getOrganizationById(id);
-        if (null != organizationByIds) {
-            List<Organization> list = organizationDAO.listTreeByCodeForParent(organizationByIds.getCode());
+        // 根据id查询到该节点信息
+        Map<Object, Object> map = new HashMap<>();
+        map.put("id", id);
+        List<Organization> organizationByIds = organizationDAO.listOrganizationBy(map);
+        if (!CollectionUtils.isEmpty(organizationByIds)) {
+            List<Organization> list = organizationDAO.listTreeByCodeForParent(organizationByIds.get(0).getCode());
             if (!CollectionUtils.isEmpty(list)) {
                 return list;
             }
@@ -198,12 +193,14 @@ public class OrganizationServiceImpl implements OrganizationService {
     @Override
     public Integer moveOrganization(String id, String parentOrgId) {
         // 根据id查询到该节点信息
-        Organization org = organizationDAO.getOrganizationById(id);
-        if (org == null) {
+        Map<Object, Object> map = new HashMap<>();
+        map.put("id", id);
+        List<Organization> org = organizationDAO.listOrganizationBy(map);
+        if (CollectionUtils.isEmpty(org)) {
             return Integer.valueOf(0);
         }
         // 根据该节点的code查询到其下所有子节点信息的集合(要移动的节点及其子节点的集合)
-        List<Organization> list = organizationDAO.listTreeByCodeForSon(org.getCode());
+        List<Organization> list = organizationDAO.listTreeByCodeForSon(org.get(0).getCode());
         /*
          * 接下来是将要移动的几点按原来的机构新增到现在的父节点上,
          * 先将该节点新增，并修改其code，parentId，his_permission；然后再新增其子节点
@@ -211,25 +208,29 @@ public class OrganizationServiceImpl implements OrganizationService {
         // 这里是移动后的code
         String code = null;
         // 先判断新的父节点下是否有子节点
-        Organization orgParent = organizationDAO.getOrganizationById(parentOrgId);
+        map = new HashMap<>();
+        map.put("id", parentOrgId);
+        List<Organization> orgParent = organizationDAO.listOrganizationBy(map);
         // 得到新父节点的子节点
-        List<Organization> listSon = organizationDAO.listByParentId(orgParent.getCode());
-        if (null != listSon && listSon.size() != 0) {
+        map = new HashMap<>();
+        map.put("parentId", orgParent.get(0).getCode());
+        List<Organization> listSon = organizationDAO.listOrganizationBy(map);
+        if (!CollectionUtils.isEmpty(listSon)) {
             List<String> codes = new ArrayList<String>();
             for (Organization orgaSon : listSon) {
                 codes.add(orgaSon.getCode());
             }
-            code = UuidUtil.getCodeByBrother(orgParent.getCode(), codes);
+            code = UuidUtil.getCodeByBrother(orgParent.get(0).getCode(), codes);
         } else {
-            code = orgParent.getCode() + "01";
+            code = orgParent.get(0).getCode() + "01";
         }
         Organization organization = new Organization();
         organization.setId(UuidUtil.getUUID());
-        organization.setOrgName(org.getOrgName());
-        organization.setuId(org.getuId());// ??????????????????????????????????????
+        organization.setOrgName(org.get(0).getOrgName());
+        organization.setuId(org.get(0).getuId());// ??????????????????????????????????????
         organization.setCode(code);
-        organization.setParentId(orgParent.getCode());
-        organization.setHis_permission(org.getHis_permission() + "," + code);
+        organization.setParentId(orgParent.get(0).getCode());
+        organization.setHis_permission(org.get(0).getHis_permission() + "," + code);
         organizationDAO.saveOrganization(organization);
 
         /*
@@ -245,8 +246,8 @@ public class OrganizationServiceImpl implements OrganizationService {
              */
             if (orga.getId() != id && !id.equals(orga.getId())) {
                 // 子节点的code
-                String code1 = orga.getCode().replaceFirst(org.getCode(), organization.getCode());
-                String parentId1 = orga.getParentId().replaceFirst(org.getCode(), organization.getCode());
+                String code1 = orga.getCode().replaceFirst(org.get(0).getCode(), organization.getCode());
+                String parentId1 = orga.getParentId().replaceFirst(org.get(0).getCode(), organization.getCode());
                 Organization organization1 = new Organization();
                 organization1.setId(UuidUtil.getUUID());
                 organization1.setOrgName(orga.getOrgName());
@@ -270,7 +271,9 @@ public class OrganizationServiceImpl implements OrganizationService {
         // 查询到当前节点
         List<Organization> organizationByIds = organizationDAO.listOrganizationBy(map);
         if (!CollectionUtils.isEmpty(organizationByIds)) {
-            List<Organization> list = organizationDAO.listByParentId(organizationByIds.get(0).getCode());
+            Map<Object, Object> mapp = new HashMap<>();
+            mapp.put("parentId", organizationByIds.get(0).getCode());
+            List<Organization> list = organizationDAO.listOrganizationBy(mapp);
             if (!CollectionUtils.isEmpty(list)) {
                 flag = true;
             }
@@ -283,9 +286,11 @@ public class OrganizationServiceImpl implements OrganizationService {
      */
     @Override
     public Organization getCompanyNameBySon(String id) {
-        Organization organizationByIds = organizationDAO.getOrganizationById(id);
-        if (null != organizationByIds) {
-            List<Organization> parent = organizationDAO.listTreeByCodeForParent(organizationByIds.getCode());
+        Map<Object, Object> map = new HashMap<Object, Object>();
+        map.put("id", id);
+        List<Organization> organizationByIds = organizationDAO.listOrganizationBy(map);
+        if (!CollectionUtils.isEmpty(organizationByIds)) {
+            List<Organization> parent = organizationDAO.listTreeByCodeForParent(organizationByIds.get(0).getCode());
             if (!CollectionUtils.isEmpty(parent)) {
                 for (Organization o : parent) {
                     if (o.getOrgName().contains("公司")) {
